requirements.txt
asgiref==3.5.0
Django==3.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-health-check==3.16.5
gunicorn==20.0.4
mysqlclient==2.0.3
Pillow==8.2.0
pytz==2022.1
sqlparse==0.4.2
